backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pillow
opencv-python-headless
numpy